feat: implement login functionality with Google and add LoginForm component
