--- src/components/SessionProvider.tsx
@@ -1,9 +1,0 @@
-'use client'
-import { SessionProvider } from "next-auth/react";
-
-export default function App({children} : {children: React.ReactNode}) {
-  return (
-    <SessionProvider >{children}
-    </SessionProvider>
-  );
-}
